# src/backend/generated/src/openapi_server/models/annotation_request.py
# coding: utf-8

"""
    3D Annotation API with JWT

    A JWT-Protected API for 3D Object Annotation

    The version of the OpenAPI document: 1.0.0
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
import pprint
import re  # noqa: F401
import json




from pydantic import BaseModel, ConfigDict, Field, StrictInt, StrictStr, field_validator
from typing import Any, ClassVar, Dict, List, Optional
from openapi_server.models.annotation_request_position import AnnotationRequestPosition
from openapi_server.models.annotation_request_size import AnnotationRequestSize
try:
    from typing import Self
except ImportError:
    from typing_extensions import Self

class AnnotationRequest(BaseModel):
    """
    AnnotationRequest
    """ # noqa: E501
    episode_id: StrictInt = Field(description="The ID of the episode to annotate")
    pcd_file_id: StrictInt = Field(description="The ID of the .pcd file to annotate")
    label: StrictStr = Field(description="Label for the annotated object")
    geometry: StrictStr = Field(description="The shape of the bounding box")
    size: Optional[AnnotationRequestSize] = None
    position: AnnotationRequestPosition
    rotation: Optional[AnnotationRequestPosition] = None
    __properties: ClassVar[List[str]] = ["episode_id", "pcd_file_id", "label", "geometry", "size", "position", "rotation"]

    @field_validator('geometry')
    def geometry_validate_enum(cls, value):
        """Validates the enum"""
        if value not in ('sphere', 'cube', 'polygon',):
            raise ValueError("must be one of enum values ('sphere', 'cube', 'polygon')")
        return value

    model_config = {
        "populate_by_name": True,
        "validate_assignment": True,
        "protected_namespaces": (),
    }


    def to_str(self) -> str:
        """Returns the string representation of the model using alias"""
        return pprint.pformat(self.model_dump(by_alias=True))

    def to_json(self) -> str:
        """Returns the JSON representation of the model using alias"""
        # TODO: pydantic v2: use .model_dump_json(by_alias=True, exclude_unset=True) instead
        return json.dumps(self.to_dict())

    @classmethod
    def from_json(cls, json_str: str) -> Self:
        """Create an instance of AnnotationRequest from a JSON string"""
        return cls.from_dict(json.loads(json_str))

    def to_dict(self) -> Dict[str, Any]:
        """Return the dictionary representation of the model using alias.

        This has the following differences from calling pydantic's
        `self.model_dump(by_alias=True)`:

        * `None` is only added to the output dict for nullable fields that
          were set at model initialization. Other fields with value `None`
          are ignored.
        """
        _dict = self.model_dump(
            by_alias=True,
            exclude={
            },
            exclude_none=True,
        )
        # override the default output from pydantic by calling `to_dict()` of size
        if self.size:
            _dict['size'] = self.size.to_dict()
        # override the default output from pydantic by calling `to_dict()` of position
        if self.position:
            _dict['position'] = self.position.to_dict()
        # override the default output from pydantic by calling `to_dict()` of rotation
        if self.rotation:
            _dict['rotation'] = self.rotation.to_dict()
        return _dict

    @classmethod
    def from_dict(cls, obj: Dict) -> Self:
        """Create an instance of AnnotationRequest from a dict"""
        if obj is None:
            return None

        if not isinstance(obj, dict):
            return cls.model_validate(obj)

        _obj = cls.model_validate({
            "episode_id": obj.get("episode_id"),
            "pcd_file_id": obj.get("pcd_file_id"),
            "label": obj.get("label"),
            "geometry": obj.get("geometry"),
            "size": AnnotationRequestSize.from_dict(obj.get("size")) if obj.get("size") is not None else None,
            "position": AnnotationRequestPosition.from_dict(obj.get("position")) if obj.get("position") is not None else None,
            "rotation": AnnotationRequestPosition.from_dict(obj.get("rotation")) if obj.get("rotation") is not None else None
        })
        return _obj